+++ lib/inno_bundle.dart
@@ -1,5 +1,5 @@
 /// inno_bundle is not designed to be imported as library within projects,
-/// Rather its a CLI tool with customized output based on pubspec.yaml configuration.
+/// Rather its a CLI tool with customized output based on its configuration.
 library;
 
 /// only empty main.