+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+## 0.10.0
+
+- **`(Breaking!)`** Change the default build type from `debug` to `release`.
+
+- Adds `--path` CLI parameter to specify custom config file path [#15](https://github.com/hahouari/inno_bundle/pull/15#issuecomment-3268840258), also support reading config from `inno_bundle.yaml` file if available.
+
+- Add PATH variable `CONFIG_FILE` to CLI output of `dart run inno_bundle:build --envs`.
+
+- Update help and error messages to reflect which config file is being used.
+
+- Change unresolved publisher name from `Unknown User` to `Unknown Publisher`.
+
 ## 0.9.0
 
 - Adds config to control VC++ redistributable, either to bundle it with the installer or download it during installation, [#13](https://github.com/hahouari/inno_bundle/issues/13)